fix: fixed explorer header
