+++ src/app/(light)/explorer/components/header/header.tsx
@@ -1,72 +1,14 @@
-import Link from 'next/link';
-import { ReactNode } from 'react';
+import { Box, BoxProps } from '@mui/material';
 
-import {
-  Box,
-  Container,
-  Link as MuiLink,
-  Stack,
-  SxProps,
-  Typography,
-} from '@mui/material';
-
-import InfoCard from '../info-card/info-card';
-
-type Props = {
-  title: string;
-  subtitle: string;
-  help: string;
-  sx?: SxProps;
-};
-
-export default function ExplorerHeader({ title, subtitle, help, sx }: Props) {
+export default function ExplorerHeader({ sx, ...props }: BoxProps) {
   return (
     <Box
+      {...props}
       sx={{
         pt: 21,
         pb: 6,
         ...sx,
       }}
-    >
-      <Stack
-        component={Container}
-        maxWidth="xl"
-        justifyContent="space-between"
-        direction="row"
-        gap={2}
-      >
-        <Box
-          sx={{
-            width: {
-              xs: '100%',
-              md: 'calc(50% - 16px)',
-              lg: '50%',
-            },
-          }}
-        >
-          <Typography component="h1" variant="h2" fontWeight="300" mb={2}>
-            {title}
-          </Typography>
-          <Typography color="text.secondary" mb={3}>
-            {subtitle}
-          </Typography>
-          <MuiLink component={Link} href="/" fontWeight="700" underline="hover">
-            {help}
-          </MuiLink>
-        </Box>
-        <InfoCard
-          sx={{
-            width: {
-              md: 'calc(50% - 16px)',
-              lg: 'calc(25% - 16px)',
-            },
-            display: {
-              xs: 'none',
-              md: 'flex',
-            },
-          }}
-        />
-      </Stack>
-    </Box>
+    />
   );
 }
