+++ src/app/(light)/explorer/components/hero/hero.tsx
@@ -1,0 +1,71 @@
+import Link from 'next/link';
+
+import {
+  Box,
+  Container,
+  Link as MuiLink,
+  Stack,
+  SxProps,
+  Typography,
+} from '@mui/material';
+
+import InfoCard from '../info-card/info-card';
+
+type Props = {
+  title: string;
+  subtitle: string;
+  help: string;
+  sx?: SxProps;
+};
+
+export default function ExplorerHero({ title, subtitle, help, sx }: Props) {
+  return (
+    <Box
+      sx={{
+        pt: 21,
+        pb: 6,
+        ...sx,
+      }}
+    >
+      <Stack
+        component={Container}
+        maxWidth="xl"
+        justifyContent="space-between"
+        direction="row"
+        gap={2}
+      >
+        <Box
+          sx={{
+            width: {
+              xs: '100%',
+              md: 'calc(50% - 16px)',
+              lg: '50%',
+            },
+          }}
+        >
+          <Typography component="h1" variant="h2" fontWeight="300" mb={2}>
+            {title}
+          </Typography>
+          <Typography color="text.secondary" mb={3}>
+            {subtitle}
+          </Typography>
+          <MuiLink component={Link} href="/" fontWeight="700" underline="hover">
+            {help}
+          </MuiLink>
+        </Box>
+        <InfoCard
+          sx={{
+            width: {
+              md: 'calc(50% - 16px)',
+              lg: 'calc(25% - 16px)',
+            },
+            display: {
+              xs: 'none',
+              md: 'flex',
+            },
+          }}
+        />
+      </Stack>
+    </Box>
+  );
+}
