+++ src/app/(light)/explorer/data-models/page.tsx
@@ -2,14 +2,14 @@ import { explorerDataModels } from '@/locale/en/datamodel';
 
 import { Divider } from '@mui/material';
 
-import ExplorerHeader from '../components/header/header';
+import ExplorerHero from '../components/hero/hero';
 import DataModelsExplorerFeatured from './components/featured';
 import DataModelsExplorerSearch from './components/search/search';
 
 export default function DataModelsExplorerPage() {
   return (
     <>
-      <ExplorerHeader
+      <ExplorerHero
         title={explorerDataModels.title}
         subtitle={explorerDataModels.subtitle}
         help={explorerDataModels.help}
